dev-0524 Moved ContextType back into types.ts to fix type declaration issues.
